decided to collect all steps of pipeline under the 'localhost:4200/model-development' domain. hence in the same scrollable page.
