--- frontend/src/app/preprocessing/preprocessing.component.html
@@ -1,21 +1,0 @@
-<div class="preprocessing-declaration">
-    <h2>Data Purifier Declaration</h2>
-  
-    <div class="selected-options">
-      <h3>Selected Options:</h3>
-      <ul>
-        <li *ngFor="let option of selectedOptions">
-          {{ option.id }}. {{ option.name }}
-        </li>
-      </ul>
-    </div>
-  
-    <mat-form-field appearance="fill">
-      <mat-label>Select Data Purification Options</mat-label>
-      <mat-select multiple [(ngModel)]="selectedOptions" (selectionChange)="onSelectionChange($event)">
-        <mat-option *ngFor="let option of purifierOptions" [value]="option" [disabled]="isOptionDisabled(option)">
-          {{ option.id }}. {{ option.name }}
-        </mat-option>
-      </mat-select>
-    </mat-form-field>
-  </div>
